refactor: update more imports
