+++ log/logger_test.ts
@@ -2,7 +2,7 @@
 import { assert, assertEquals, assertMatch } from "../assert/mod.ts";
 import { Logger, LogRecord } from "./logger.ts";
 import { LevelName, LogLevels } from "./levels.ts";
-import { BaseHandler } from "./handlers.ts";
+import { BaseHandler } from "./base_handler.ts";
 
 class TestHandler extends BaseHandler {
   public messages: string[] = [];
